requirements.txt
datetime
pandas==2.1.1
request
json
os
awswrangler